2024-10-02-leap-CUDA: make driver installation instructions more robust
Open kernel driver KMP must be in sync with userspace drivers.
Therefore it is recommended to install the KMP first, then
figure out which version has been installed, and then install
the corresponding user space drivers.

On Bare Metal it was missing the installation of the userspace drivers
completely.

CUDA is meanwhile at version 12.9 with 575 drivers.

-32bit packages are no longer recommended. Therefore --no-recommends
is no longer needed.

Unfortunately we always need to specify the version when installing
userspace drivers (drivers need to be in sync; see above).